[debian patch] 07_gallium-fix-build-failure-on-powerpcspe.diff
gallium: fix build failure on powerpcspe

In the case of powerpc, mesa activates some altivec instructions
that are unknown on the powerpcspe architecture (see
https://wiki.debian.org/PowerPCSPEPort), causing a build failure as the
'vand' opcode is not recognized by the assembler.

This patch fixes this by preventing the PPC-specialcasing in case of
powerpcspe (__NO_FPRS__ is only defined there).

https://bugs.debian.org/695746
Signed-off-by: Han Gao <gaohan@iscas.ac.cn>